basic configuration, UUID, and registration support
Adds the ConfigManager() class that allows for configuration settings via CLI,
environmental variables, a configuration file (kintyre_speedtest.ini), and
various built-in defaults.

Created a new "--register" mode that allows for the initial population of
endpoint settings and the local agent's UUID.  This UUID is now sent in as the
'host' value instead of OS' hostname (which can change based on network
configuration, when switching between DHCP servers)  When registration occurs,
a "registration" event is sent to the HEC endpoint to test connectivity.
Eventually the registration mode will be setup to prompt the user to provide
necessary values, but for now the user must supply them as CLI args or
environmental variables. (Closes #12)

The UUID generation support closes #11.

Updated the README.md docs to cleanup the install steps and to show a usage
example of the new register function (which now must be run manually once,
since the HEC endpoint and token are no longer hard-coded in the script.)

Travis and Tox configurations were updated to include status UUID values and
names that should allow for better identification (and filtering) of automation
testing-based events from Splunk.