Add interactive registration mode
Revamped registration process to allow both interactive and non-interactive
modes.  Interactive mode is now the default, which will cause the user to be
prompted for select agent and endpoint configuration settings.  Non-interactive
mode can be entered by using the '--no-prompt' CLI option in which case any
CLI or environmental variables set will be pushed into the config file.

Send a ping on startup and abort on failure.  Closes #13.

Tox and Travis testing have been updated with the '--no-prompt' option.

The ConfigOption class was updated to support the 'prompt' attribute to indicate
which fields should be prompted for interactively.  Also ini_* helper properties
were added for the new find_options() method.

The ConfigManager was updated to revamp the handling of various set/get methods
and attribute access.  The items() method was added to iterate over the known
configuration and find_options() allows options to be searched based on various
attribute=value pairs.

The tox configuration was updated to allow passing for the passing of positional
arguments to the CLI (so that '--fake-it' can be passed in to accelerate various
testing cycles.)  Also the agent config file is now stored in the local .tox
folder instead of being stored per environment, thus allowing for a random yet
persistent UUID.